+++ crates/ruma-common/src/canonical_json/value.rs
@@ -178,10 +178,21 @@ impl TryFrom<JsonValue> for CanonicalJsonValue {
     fn try_from(val: JsonValue) -> Result<Self, Self::Error> {
         Ok(match val {
             JsonValue::Bool(b) => Self::Bool(b),
-            JsonValue::Number(num) => Self::Integer(
-                Int::try_from(num.as_i64().ok_or(CanonicalJsonError::IntConvert)?)
-                    .map_err(|_| CanonicalJsonError::IntConvert)?,
-            ),
+            JsonValue::Number(num) => {
+                // Treat float separately to get a better error message.
+                //
+                // We do several checks because the docs say that `num.is_f64()` is not guaranteed
+                // to return `false` for an integer in the future.
+                if !num.is_i64() && !num.is_u64() && num.is_f64() {
+                    return Err(CanonicalJsonError::InvalidType("float".to_owned()));
+                }
+
+                Self::Integer(
+                    num.as_i64()
+                        .and_then(|num| Int::try_from(num).ok())
+                        .ok_or(CanonicalJsonError::IntegerOutOfRange)?,
+                )
+            }
             JsonValue::Array(vec) => {
                 Self::Array(vec.into_iter().map(TryInto::try_into).collect::<Result<Vec<_>, _>>()?)
             }