+++ crates/ruma-common/CHANGELOG.md
@@ -1,5 +1,21 @@
 # [unreleased]
 
+Breaking changes:
+
+- The variants of `CanonicalJsonError` variants changed.
+- `to_canonical_value()` has stricter rules, it now returns errors for the
+  following cases which were never documented to work in the first place:
+  - Serializing bytes.
+  - Serializing booleans and integers as keys for an object.
+  - Serializing the same key twice in an object.
+
+Improvements:
+
+- Add `canonical_json::Serializer`, which allows to serialize a type directly to
+  a `CanonicalJsonValue`, with stricter rules than
+  `serde_json::value::Serializer`. This serializer is also used in
+  `to_canonical_value()`.
+
 # 0.17.1
 
 Bug fixes: